Sort out extensions by groups
Also:
 * Rename `AstExtensions.kt` to `AstExts.kt` to according to latest conventions.
