+++ api/src/main/kotlin/io/spine/protodata/DescriptorToAst.kt
